Move model parameters declaration to :toolingApi
